Randomization inference
Perform 100 simulations where 100 additional simulations are conducted inside each of the overall repetitions. In the 'inner' simulations, treatment assignment at the village level is re-calculated. The 100 betas are saved and compared to the main one (in the 100 overall regressions). The percent of betas >= the main one give the one-tailed p-value. This is the statistic that comes out of the simulations. So for each specification in the loop, 100 one-tailed and two-tailed p-values are saved. The two-tailed is the same as the one-tailed but using the abs() of the 100 inner betas